third_party/WebKit/LayoutTests/vr/requestPresent_reject_nowebgl.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/public/mojom/vr_service.mojom.js"></script>
<script src="resources/mock-vr-service.js"></script>
<canvas id="webgl-canvas"></canvas>
<canvas id="canvas-2d"></canvas>
<script src="resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test( (t) => {
  return navigator.getVRDisplays().then( (displays) => {
    var canvas2d = document.getElementById("canvas-2d");
    var ctx = canvas2d.getContext("2d");
    var display = displays[0];

    runWithUserGesture( () => {
      display.requestPresent([{ source : canvas2d }]).then( () => {
        t.step( () => {
          assert_unreached();
        }, "Display should not be presenting");
      }, (err) => {
        t.step( () => {
          assert_false(display.isPresenting);
        }, "requestPresent rejected and not presenting");
      }).then( () => {
        t.done();
      });
    });
  }, (err) => {
    t.step( () => {
      assert_unreached(err);
    }, "getVRDisplays rejected");
    t.done();
  });
}, [fakeDisplays["Pixel"]],
"Test requestPresent rejects if provided a non-webgl canvas");

</script>
